requirements.txt
bs4
discord.py
requests
datetime
wikipedia
memory_profiler
googletrans
selenium